feat: single-header and const-correctness
Make library single-header (with option to more complex use-cases) and
improve const-correctness
